+++ src/main/java/com/pingplusplus/model/SettleAccountRecipient.java
@@ -10,6 +10,11 @@ public class SettleAccountRecipient extends PingppObject {
     String openBankCode;
     String openBank;
     Boolean forceCheck;
+    String accountType;
+    Integer cardType;
+    String prov;
+    String city;
+    String subBank;
 
     public String getAccount() {
         return account;
@@ -57,5 +62,45 @@ public class SettleAccountRecipient extends PingppObject {
 
     public void setForceCheck(Boolean forceCheck) {
         this.forceCheck = forceCheck;
+    }
+
+    public String getAccountType() {
+        return accountType;
+    }
+
+    public void setAccountType(String accountType) {
+        this.accountType = accountType;
+    }
+
+    public Integer getCardType() {
+        return cardType;
+    }
+
+    public void setCardType(Integer cardType) {
+        this.cardType = cardType;
+    }
+
+    public String getProv() {
+        return prov;
+    }
+
+    public void setProv(String prov) {
+        this.prov = prov;
+    }
+
+    public String getCity() {
+        return city;
+    }
+
+    public void setCity(String city) {
+        this.city = city;
+    }
+
+    public String getSubBank() {
+        return subBank;
+    }
+
+    public void setSubBank(String subBank) {
+        this.subBank = subBank;
     }
 }