+++ src/test/java/com/pingplusplus/PingppAccountTest.java
@@ -17,7 +17,7 @@ public class PingppAccountTest extends PingppAccountTestBase {
     }
 
     @Test public void testVerifyVersions() {
-        assertEquals("Pingpp.VERSION should match", "2.3.3", Pingpp.VERSION);
+        assertEquals("Pingpp.VERSION should match", "2.3.4", Pingpp.VERSION);
         assertEquals("PingppAccount.VERSION should match", "1.2.4", PingppAccount.VERSION);
     }
 }